refactor(desktop): 拆出 appearance/networks/dream/developer 设置面板
将四个低耦合设置面板迁入 settings/panels/，llmHttpVersion 选项并入 constants。
