+++ apps/desktop/src/components/settings-view.tsx
@@ -118,7 +118,11 @@ import {
   hasGoogleVertexServiceAccountCredentials,
 } from "@/host/provider-api-key";
 import { AgentsSettingsPanel } from "@/components/settings/panels/agents-settings-panel";
-import { settingsPageTitleKey, themeSelectOptions } from "@/components/settings/constants";
+import { AppearanceSettingsPanel } from "@/components/settings/panels/appearance-settings-panel";
+import { DeveloperSettingsPanel } from "@/components/settings/panels/developer-settings-panel";
+import { DreamSettingsPanel } from "@/components/settings/panels/dream-settings-panel";
+import { NetworksSettingsPanel } from "@/components/settings/panels/networks-settings-panel";
+import { settingsPageTitleKey } from "@/components/settings/constants";
 import { formatExtensionInstalledAt, fileToBase64, formatSettingsTime } from "@/components/settings/formatters";
 import { SettingsRow } from "@/components/settings/settings-row";
 import type { SettingsFormState, SettingsViewProps } from "@/components/settings/types";
@@ -597,66 +601,6 @@ function skillRootKindLabel(rootKind: DesktopSkillRootKind): string {
 
 function skillLocationLabel(item: DesktopSkillListItem): string {
   return skillRootKindLabel(item.rootKind);
-}
-
-function webHostStatusLabel(state: DesktopSnapshot["webHost"]["status"]["state"]): string {
-  switch (state) {
-    case "running":
-      return i18n.t('settings.webHostRunning');
-    case "starting":
-      return i18n.t('settings.webHostStarting');
-    case "error":
-      return i18n.t('settings.webHostError');
-    case "stopped":
-      return i18n.t('settings.webHostStopped');
-    default:
-      return i18n.t('settings.webHostClosed');
-  }
-}
-
-function dreamCollectorStateLabel(state: DesktopSnapshot["dreams"]["collector"]["state"]): string {
-  switch (state) {
-    case "disabled":
-      return i18n.t('settings.dreamDisabled');
-    case "missing-model":
-      return i18n.t('settings.dreamMissingModel');
-    case "running":
-      return i18n.t('settings.dreamCollecting');
-    case "backoff":
-      return i18n.t('settings.dreamBackoff');
-    case "error":
-      return i18n.t('settings.dreamError');
-    default:
-      return i18n.t('settings.dreamIdle');
-  }
-}
-
-function DeveloperSettingsPanel({
-  onStartCompactionUiDemo,
-}: Pick<SettingsViewProps, "onStartCompactionUiDemo">) {
-  const { t } = useTranslation();
-  return (
-    <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80 px-4 sm:px-5">
-      <div className="flex flex-col gap-2 py-4 sm:flex-row sm:items-center sm:justify-between">
-        <div className="min-w-0 space-y-1">
-          <p className="text-sm font-medium text-foreground">{t('settings.compactionDemoTitle')}</p>
-          <p className="text-xs leading-5 text-muted-foreground">
-            {t('settings.compactionDemoDescription')}
-          </p>
-        </div>
-        <Button
-          type="button"
-          variant="outline"
-          size="sm"
-          className="shrink-0 self-start sm:self-center"
-          disabled={!onStartCompactionUiDemo}
-          onClick={() => onStartCompactionUiDemo?.()}
-        >
-          {t('settings.demoInConversation')}
-        </Button>
-      </div>
-    </div>
-  );
 }
 
 const skillCreateRootOptions: Array<{
@@ -4194,399 +4138,6 @@ function ModelsSettingsPanel({
           </div>
         </DialogContent>
       </Dialog>
-    </div>
-  );
-}
-
-const llmHttpVersionSelectOptions = [
-  { value: "http1.1" as const, labelKey: "settings.llmHttpVersionHttp11" },
-  { value: "http2" as const, labelKey: "settings.llmHttpVersionHttp2" },
-];
-
-function NetworksSettingsPanel({
-  settings,
-  snapshot,
-  onSavePatch,
-  onResetWebHostPairing,
-}: Pick<
-  SettingsViewProps,
-  "settings" | "snapshot" | "onSavePatch" | "onResetWebHostPairing"
->) {
-  const { t } = useTranslation();
-  const webHost = snapshot?.webHost;
-  const webHostUrl =
-    webHost?.status.url ?? `http://${settings.webHostHost}:${settings.webHostPort}`;
-  const webHostStatus = webHostStatusLabel(webHost?.status.state ?? "disabled");
-
-  const [webHostHostDraft, setWebHostHostDraft] = useState(settings.webHostHost);
-  const [webHostPortDraft, setWebHostPortDraft] = useState(String(settings.webHostPort));
-
-  useEffect(() => {
-    setWebHostHostDraft(settings.webHostHost);
-  }, [settings.webHostHost]);
-
-  useEffect(() => {
-    setWebHostPortDraft(String(settings.webHostPort));
-  }, [settings.webHostPort]);
-
-  return (
-    <div className="space-y-6">
-      <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80 px-4 sm:px-5">
-        <SettingsRow
-          label={t('settings.llmHttpVersion')}
-          description={t('settings.llmHttpVersionDescription')}
-          htmlFor="settings-llm-http-version-select"
-        >
-          <Select
-            value={settings.llmHttpVersion}
-            onValueChange={(value) => void onSavePatch({ llmHttpVersion: value as 'http1.1' | 'http2' })}
-          >
-            <SelectTrigger id="settings-llm-http-version-select" className="w-full sm:min-w-[12rem]">
-              <SelectValue placeholder={t('settings.llmHttpVersion')} />
-            </SelectTrigger>
-            <SelectContent>
-              {llmHttpVersionSelectOptions.map((opt) => (
-                <SelectItem key={opt.value} value={opt.value}>
-                  {t(opt.labelKey)}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        </SettingsRow>
-      </div>
-
-      <div className="space-y-3">
-        <p className="text-xs text-muted-foreground">{t('settings.remoteAccessSection')}</p>
-        <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80 px-4 sm:px-5">
-          <SettingsRow
-            label={t('settings.webRemoteAccess')}
-            description={t('settings.webRemoteAccessDescription')}
-            htmlFor="settings-web-host-enabled"
-          >
-            <div className="flex items-center justify-end gap-3">
-              <span className="truncate text-sm text-muted-foreground">
-                {settings.webHostEnabled ? webHostStatus : t('settings.webHostClosed')}
-              </span>
-              <Checkbox
-                id="settings-web-host-enabled"
-                checked={settings.webHostEnabled}
-                onCheckedChange={(value) =>
-                  void onSavePatch({ webHostEnabled: value === true })
-                }
-                className="size-5"
-              />
-            </div>
-          </SettingsRow>
-
-          <SettingsRow
-            label={t('settings.listenAddress')}
-            description={t('settings.listenAddressDescription')}
-            htmlFor="settings-web-host-host"
-          >
-            <DesktopFormInput
-              id="settings-web-host-host"
-              className="sm:text-right"
-              value={webHostHostDraft}
-              onChange={(event) => setWebHostHostDraft(event.target.value)}
-              onBlur={() => {
-                const next = webHostHostDraft.trim();
-                if (next && next !== settings.webHostHost) {
-                  void onSavePatch({ webHostHost: next });
-                }
-              }}
-              disabled={!settings.webHostEnabled}
-              placeholder="127.0.0.1"
-            />
-          </SettingsRow>
-
-          <SettingsRow label={t('settings.listenPort')} htmlFor="settings-web-host-port">
-            <DesktopFormInput
-              id="settings-web-host-port"
-              className="sm:text-right"
-              type="number"
-              min={1}
-              max={65535}
-              value={webHostPortDraft}
-              onChange={(event) => setWebHostPortDraft(event.target.value)}
-              onBlur={() => {
-                const port = Number.parseInt(webHostPortDraft, 10);
-                if (Number.isInteger(port) && port >= 1 && port <= 65535 && port !== settings.webHostPort) {
-                  void onSavePatch({ webHostPort: port });
-                }
-              }}
-              disabled={!settings.webHostEnabled}
-              placeholder="7788"
-            />
-          </SettingsRow>
-
-          <div className="py-4">
-            <p className="text-sm font-medium text-foreground">{t('settings.remoteStatus')}</p>
-            <div className="mt-2 grid gap-1 text-sm text-muted-foreground sm:text-right">
-              <p className="truncate">
-                <span className="text-foreground">{settings.webHostEnabled ? webHostStatus : t('settings.webHostClosed')}</span>
-                {settings.webHostEnabled ? ` · ${webHostUrl}` : null}
-              </p>
-              {webHost?.status.error ? (
-                <p className="break-words text-destructive">{webHost.status.error}</p>
-              ) : null}
-              <p>
-                {t('settings.pairing')}{webHost?.config.paired ? t('settings.pairingDone') : t('settings.pairingPending')}
-              </p>
-              {webHost?.status.pairingCode ? (
-                <p className="font-mono text-foreground">{webHost.status.pairingCode}</p>
-              ) : null}
-              {settings.webHostEnabled && webHost?.config.paired && onResetWebHostPairing ? (
-                <div className="mt-3 flex justify-end">
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="sm"
-                    className="text-muted-foreground"
-                    onClick={() => void onResetWebHostPairing()}
-                  >
-                    {t('settings.resetPairing')}
-                  </Button>
-                </div>
-              ) : null}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AppearanceSettingsPanel({
-  theme,
-  onThemeChange,
-  font,
-  onFontChange,
-  clickablePointerCursor,
-  onClickablePointerCursorChange,
-  settings,
-  onSavePatch,
-}: Pick<
-  SettingsViewProps,
-  | "theme"
-  | "onThemeChange"
-  | "font"
-  | "onFontChange"
-  | "clickablePointerCursor"
-  | "onClickablePointerCursorChange"
-  | "settings"
-  | "onSavePatch"
->) {
-  const { t } = useTranslation();
-  return (
-    <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80 px-4 sm:px-5">
-      <SettingsRow
-        label={t('settings.theme')}
-        description={t('settings.themeDescription')}
-        htmlFor="settings-theme-select"
-      >
-        <Select value={theme} onValueChange={(v) => onThemeChange(v as ThemePreference)}>
-          <SelectTrigger id="settings-theme-select" className="w-full sm:min-w-[12rem]">
-            <SelectValue placeholder={t('settings.selectTheme')} />
-          </SelectTrigger>
-          <SelectContent>
-            {themeSelectOptions.map((opt) => (
-              <SelectItem key={opt.value} value={opt.value}>
-                {t(opt.labelKey)}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-
-      <SettingsRow
-        label={t('settings.font')}
-        description={t('settings.fontDescription')}
-        htmlFor="settings-font-select"
-      >
-        <FontSelect id="settings-font-select" value={font} onValueChange={onFontChange} />
-      </SettingsRow>
-
-      <SettingsRow
-        label={t('settings.uiLocale')}
-        description={t('settings.uiLocaleDescription')}
-        htmlFor="settings-locale"
-      >
-        <Select
-          value={settings.uiLocale}
-          onValueChange={(value) => {
-            void changeLanguage(value);
-            void onSavePatch({ uiLocale: value });
-          }}
-        >
-          <SelectTrigger id="settings-locale" className="w-full sm:min-w-[12rem]">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            {VALID_LANGUAGES.map((lang) => (
-              <SelectItem key={lang} value={lang}>
-                {lang === 'zh-CN' ? t('settings.langZhCN') : t('settings.langEn')}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-
-      <SettingsRow
-        label={t('settings.blurEffect')}
-        description={
-          isNativeBackdropBlurSupported()
-            ? t('settings.blurEffectDescription')
-            : t('settings.blurEffectUnsupported')
-        }
-        htmlFor="settings-blur-effect"
-      >
-        {isNativeBackdropBlurSupported() ? (
-          <div className="flex justify-end">
-            <Checkbox
-              id="settings-blur-effect"
-              checked={settings.windowsMica}
-              onCheckedChange={(value) => void onSavePatch({ windowsMica: value === true })}
-              className="size-5"
-            />
-          </div>
-        ) : (
-          <p className="text-sm text-muted-foreground sm:text-right">—</p>
-        )}
-      </SettingsRow>
-
-      <SettingsRow
-        label={t('settings.clickablePointerCursor')}
-        description={t('settings.clickablePointerCursorDescription')}
-        htmlFor="settings-clickable-pointer-cursor"
-      >
-        <div className="flex justify-end">
-          <Checkbox
-            id="settings-clickable-pointer-cursor"
-            checked={clickablePointerCursor}
-            onCheckedChange={(value) => onClickablePointerCursorChange(value === true)}
-            className="size-5"
-          />
-        </div>
-      </SettingsRow>
-    </div>
-  );
-}
-
-function DreamSettingsPanel({
-  theme,
-  settings,
-  snapshot,
-  onSavePatch,
-  onListDreamsOverview,
-}: Pick<SettingsViewProps, "theme" | "settings" | "snapshot" | "onSavePatch" | "onListDreamsOverview">) {
-  const { t } = useTranslation();
-  const collector = snapshot?.dreams.collector;
-  const disabled = !settings.dreamEnabled;
-  const [dreamItems, setDreamItems] = useState<DesktopDreamOverviewItem[]>([]);
-  const [dreamsLoading, setDreamsLoading] = useState(false);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    const loadDreams = async () => {
-      setDreamsLoading(true);
-      try {
-        const items = await onListDreamsOverview();
-        if (!cancelled) {
-          setDreamItems(items);
-        }
-      } catch {
-        if (!cancelled) {
-          setDreamItems([]);
-        }
-      } finally {
-        if (!cancelled) {
-          setDreamsLoading(false);
-        }
-      }
-    };
-
-    void loadDreams();
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    onListDreamsOverview,
-    snapshot?.workspaceRoot,
-    snapshot?.git.branch,
-    snapshot?.dreams.collector.processedCount,
-    snapshot?.dreams.collector.lastSuccessAtUnixMs,
-  ]);
-
-  return (
-    <div className="space-y-6">
-      <DreamGraphCard
-        items={dreamItems}
-        loading={dreamsLoading}
-        theme={theme}
-        workspaceRoot={snapshot?.workspaceRoot}
-        gitBranch={snapshot?.git.branch}
-        collectorState={collector?.state ?? "disabled"}
-        dreamEnabled={settings.dreamEnabled}
-        debugMode={settings.dreamDebugMode}
-      />
-
-      <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80 px-4 sm:px-5">
-        <SettingsRow
-          label={t('settings.dreams')}
-          description={t('settings.dreamDescription')}
-          htmlFor="settings-dream-enabled"
-        >
-          <div className="flex items-center justify-end gap-3">
-            <Badge variant="outline">Beta</Badge>
-            <Checkbox
-              id="settings-dream-enabled"
-              checked={settings.dreamEnabled}
-              onCheckedChange={(value) => void onSavePatch({ dreamEnabled: value === true })}
-              className="size-5"
-            />
-          </div>
-        </SettingsRow>
-
-        <SettingsRow
-          label={t('settings.debugMode')}
-          description={t('settings.debugModeDescription')}
-          htmlFor="settings-dream-debug"
-        >
-          <div className="flex justify-end">
-            <Checkbox
-              id="settings-dream-debug"
-              checked={settings.dreamDebugMode}
-              disabled={disabled}
-              onCheckedChange={(value) => void onSavePatch({ dreamDebugMode: value === true })}
-              className="size-5"
-            />
-          </div>
-        </SettingsRow>
-
-        <div className="py-4">
-          <p className="text-sm font-medium text-foreground">{t('settings.collectorStatus')}</p>
-          <div className="mt-2 grid gap-1 text-sm text-muted-foreground sm:text-right">
-            <p>
-              {t('settings.status')}
-              <span className="font-medium text-foreground">
-                {dreamCollectorStateLabel(collector?.state ?? "disabled")}
-              </span>
-            </p>
-            <p>
-              {t('settings.pendingProcessed', { pending: collector?.pendingCount ?? 0, processed: collector?.processedCount ?? 0 })}
-            </p>
-            <p>{t('settings.lastRun')}{formatSettingsTime(collector?.lastRunAtUnixMs)}</p>
-            <p>{t('settings.lastSuccess')}{formatSettingsTime(collector?.lastSuccessAtUnixMs)}</p>
-            {collector?.backoffUntilUnixMs ? (
-              <p>{t('settings.backoffUntil')}{formatSettingsTime(collector.backoffUntilUnixMs)}</p>
-            ) : null}
-            {collector?.lastError ? (
-              <p className="break-words text-destructive">{collector.lastError}</p>
-            ) : null}
-          </div>
-        </div>
-      </div>
     </div>
   );
 }
